00_Serverless/04.16_SLS/requirements.txt
fastapi
sqlalchemy
uvicorn[standard]
psycopg2-binary
python-dotenv
# requests
# argparse
# colorlog
